src/requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1600699246375/work
chardet==3.0.4
click==8.0.1
colorama==0.4.4
conda==4.10.3
conda-package-handling @ file:///C:/ci/conda-package-handling_1603003327818/work
cryptography @ file:///C:/ci/cryptography_1601046905460/work
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1596215850360/work
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pdftotext @ file:///D:/bld/pdftotext_1621194315211/work
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.18
pycodestyle==2.7.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydocstyle==6.1.1
Pygments==2.8.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-docx==0.8.11
pytz==2021.1
pywin32==227
pywinpty==0.5.7
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
ruamel-yaml==0.15.87
Send2Trash==1.5.0
six==1.15.0
snowballstemmer==2.1.0
sudo==1.0.0
terminado==0.9.4
testpath==0.4.4
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1602185206534/work
traitlets==5.0.5
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
